refactor(recipes): MoE load-balance helpers → BaseRecipe (enable in VLM)
_collect_moe_load_balance + _log_moe_metrics move to BaseRecipe and are
now called from VLM's training loop too. When training MoE VLMs (e.g.
Qwen3-VL-MoE-30B) with ``moe_metrics.enabled: true`` in YAML, the wandb
/ comet trackers now receive the same load-balance summaries as LLM MoE.

VLM's log_train_metrics shim also gains mlflow_logger and comet_logger
pass-through (was wandb-only).
